Consistently use "unsigned int" for null count
Currently these types are used for null_count (and also uint8_t for
storage which is not subject to the change here):
int
unsigned int
size_t

Also use format %u for it.

+++ link-grammar/api-structures.h
@@ -162,7 +162,7 @@ struct Sentence_s
 	                               This is not the same as num alloced
 	                               because some may be non-canonical. */
 	size_t num_valid_linkages;  /* Number with no pp violations */
-	size_t null_count;          /* Number of null links in linkages */
+	unsigned int null_count;    /* Number of null links in linkages */
 	Linkage        lnkages;     /* Sorted array of valid & invalid linkages */
 	Postprocessor * postprocessor;
 	Postprocessor * constituent_pp;
